medical-ai-assistant/requirements.txt
# Core AI/ML Dependencies
google-generativeai>=0.4.0
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-qdrant

# Environment Management
python-dotenv==1.0.1

# Document Processing
pypdf

# Vector Database
qdrant-client

# Async Support
aiofiles

# Optional: Web Framework (for future API development)
fastapi
uvicorn 